pyproject.toml
[tool.poetry]
name = "autoparts"
version = "0.1.0"
description = ""
authors = ["João Victor Inocêncio Macedo de Aguiar <macdjoao@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.21.0"
sqlalchemy = "^1.4.45"
alembic = "^1.8.1"
sqlacodegen = "^2.3.0"
mysql-connector-python = "^8.0.31"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
